requirements.txt
# Django Core
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Image Processing & Storage
Pillow>=10.1.0
django-cloudinary-storage>=0.3.0
cloudinary>=1.36.0

# API Framework
djangorestframework>=3.14.0
markdown>=3.4.4  # Browsable API support
django-filter>=23.2  # Filtering support
djangorestframework-simplejwt>=5.3.0  # JWT Authentication

# Environment Variables
python-dotenv>=1.0.0

# Security
django-cors-headers>=4.3.0

# For fly.io deployment
django-environ>=0.11.2

# Export/Import Features
openpyxl>=3.1.2
reportlab>=4.0.0
xlrd>=2.0.1
weasyprint>=60.0  # PDF generation

# Background Tasks (for scheduled reminders)
django-q2>=1.6.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0

# Additional utilities
python-dateutil>=2.8.2

